NOC/nocproject/netbox/ip_scan/ip_scan/requirements.txt

aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
bleach==6.0.0
boto3==1.26.156
botocore==1.29.158
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
Django==4.1.9
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-graphiql-debug-toolbar==0.2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-pglocks==1.0.4
django-prometheus==2.3.1
django-redis==5.3.0
django-rich==1.6.0
django-rq==2.8.1
django-storages==1.13.2
django-tables2==2.5.3
django-taggit==4.0.0
django-timezone-field==5.1
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
dulwich==0.21.5
ecdsa==0.18.0
feedparser==6.0.10
future==0.18.3
ghp-import==2.1.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
ip-scan @ file:///opt/netbox/netbox/extras/plugins/ip_scan
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
junos-eznc==2.6.7
lxml==4.9.3
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.4.3
mkdocs-autorefs==0.4.1
mkdocs-material==9.1.16
mkdocs-material-extensions==1.1.1
mkdocstrings==0.22.0
mkdocstrings-python-legacy==0.2.3
ncclient==0.6.13
netaddr==0.8.0
netbox-interface-sync==0.2.0
netmiko==4.2.0
ntc-templates==3.5.0
oauthlib==3.2.2
packaging==23.1
paramiko==3.2.0
phonebox-plugin==0.0.6b1
Pillow==9.5.0
pkg_resources==0.0.0
prometheus-client==0.17.0
promise==2.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pymdown-extensions==10.0.1
PyNaCl==1.5.0
pynetbox==7.0.1
pyparsing==3.1.1
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytkdocs==0.16.1
pytz==2023.3
PyYAML==6.0
pyyaml_env_tag==0.1
redis==4.5.5
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rq==1.15.1
rsa==4.9
s3transfer==0.6.1
scp==0.14.5
sentry-sdk==1.25.1
sgmllib3k==1.0.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
svgwrite==1.4.3
tablib==3.5.0
test-plagin @ file:///opt/netbox/netbox/extras/plugins/test_plagin
text-unidecode==1.3
textfsm==1.1.3
transitions==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
watchdog==3.0.0
webencodings==0.5.1
yamlordereddictloader==0.4.0
zipp==3.15.0
